Rename "hire" to "contact"
