+++ src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 
-import Hire from './Pages/Hire';
+import Contact from './Pages/Contact';
 import Home from './Pages/Home';
 import NotFound from './Pages/NotFound';
 import Projects from './Pages/Projects';
@@ -18,7 +18,7 @@ class App extends Component {
           <Route path="/projects" component={Projects} />
           <Route path="/writing/:slug" component={Detail} />
           <Route path="/writing" component={Writing} />
-          <Route path="/hire" component={Hire} />
+          <Route path="/contact" component={Contact} />
           <Route path="/" component={NotFound} />
         </Switch>
       </Router>
